MGDAPI-5728 refactor alert manager secret creation
MGDAPI-5728 update unit tests

MGDAPI-5728 address comments
