+++ pkg/products/threescale/reconciler.go
@@ -14,7 +14,7 @@ import (
 	portaClient "github.com/3scale/3scale-porta-go-client/client"
 	"github.com/integr8ly/integreatly-operator/pkg/addon"
 	"github.com/integr8ly/integreatly-operator/pkg/resources/k8s"
-	operatorsv1alpha1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/apis/operators/v1alpha1"
+	operatorsv1alpha1 "github.com/operator-framework/api/pkg/operators/v1alpha1"
 	k8sTypes "k8s.io/apimachinery/pkg/types"
 
 	"github.com/golang/protobuf/ptypes/any"
@@ -1459,12 +1459,7 @@ func (r *Reconciler) reconcileExternalDatasources(ctx context.Context, serverCli
 }
 
 func (r *Reconciler) reconcileOutgoingEmailAddress(ctx context.Context, serverClient k8sclient.Client) (integreatlyv1alpha1.StatusPhase, error) {
-	observabilityConfig, err := r.ConfigManager.ReadObservability()
-	if err != nil {
-		return integreatlyv1alpha1.PhaseFailed, err
-	}
-
-	existingSMTPFromAddress, err := resources.GetSMTPFromAddress(ctx, serverClient, r.log, r.installation, observabilityConfig.GetNamespace())
+	existingSMTPFromAddress, err := resources.GetSMTPFromAddress(ctx, serverClient, r.log, r.installation)
 	if err != nil {
 		return integreatlyv1alpha1.PhaseFailed, err
 	}
@@ -3722,12 +3717,7 @@ func (r *Reconciler) reconcileServiceMonitor(ctx context.Context, client k8sclie
 }
 
 func (r *Reconciler) reconcileDcEnvarEmailAddress(ctx context.Context, serverClient k8sclient.Client, dcName string, updateFn func(dc *appsv1.DeploymentConfig, value string) bool) (integreatlyv1alpha1.StatusPhase, error) {
-	observabilityConfig, err := r.ConfigManager.ReadObservability()
-	if err != nil {
-		return integreatlyv1alpha1.PhaseFailed, err
-	}
-
-	existingSMTPFromAddress, err := resources.GetSMTPFromAddress(ctx, serverClient, r.log, r.installation, observabilityConfig.GetNamespace())
+	existingSMTPFromAddress, err := resources.GetSMTPFromAddress(ctx, serverClient, r.log, r.installation)
 	if err != nil {
 		return integreatlyv1alpha1.PhaseFailed, err
 	}
@@ -3756,12 +3746,7 @@ func (r *Reconciler) reconcileDcEnvarEmailAddress(ctx context.Context, serverCli
 }
 
 func (r *Reconciler) syncInvitationEmail(ctx context.Context, serverClient k8sclient.Client) (integreatlyv1alpha1.StatusPhase, error) {
-	observabilityConfig, err := r.ConfigManager.ReadObservability()
-	if err != nil {
-		return integreatlyv1alpha1.PhaseFailed, err
-	}
-
-	fromAddress, err := resources.GetSMTPFromAddress(ctx, serverClient, r.log, r.installation, observabilityConfig.GetNamespace())
+	fromAddress, err := resources.GetSMTPFromAddress(ctx, serverClient, r.log, r.installation)
 	if err != nil {
 		return integreatlyv1alpha1.PhaseFailed, err
 	}
